Move Scope to its own file and use dict of dict not OrderedDict
Dicts are now ordered in all officially supported versions of python as 3.6 has reached end of life
